Add methods in aux methods file
- New methods for forms to reduce duplicate code
